refactor(renderer): isolate campus-data module on 2.0.2 base [skip ci]
