+++ desktop/renderer/features/campus-data/index.mjs
@@ -1,0 +1,2 @@
+export { create } from './controller.mjs';
+export { scheduleWeekModel, scheduleWeekLayout, weekRange } from './calendar-model.mjs';
