+++ desktop/renderer/index.html
@@ -733,12 +733,11 @@
   <script src="official-favorite-dialog.js"></script>
   <script src="add-website-dialog.js"></script>
   <script src="campus-service-workspace.js"></script>
-  <script src="campus-data-modules.js"></script>
   <script src="proxy-auth-migration.js"></script>
   <script src="school-profile-selector.js"></script>
   <script src="integration-center.js"></script>
-  <script src="app.js"></script>
-  <script src="auth-challenge.js"></script>
+  <script type="module" src="app.js"></script>
+  <script type="module" src="auth-challenge.js"></script>
   <p id="cardBoardLiveRegion" class="visually-hidden" aria-live="polite"></p>
 </body>
 </html>
