+++ desktop/test/javascript-syntax-gate.test.js
@@ -3,11 +3,24 @@
 const assert = require('node:assert/strict');
 const test = require('node:test');
 const {
+  controlModuleEntrypoints,
   checkJavaScriptSource,
   listJavaScriptFiles,
   parseArguments,
   safeTrackedPath,
 } = require('../scripts/check-javascript-syntax');
+
+test('HTML module declarations select module parsing independently of the js extension', () => {
+  assert.deepEqual([...controlModuleEntrypoints(`
+    <!-- <script type="module" src="ignored.js"></script> -->
+    <script src="classic.js"></script>
+    <script type="module" src="app.js"></script>
+    <script src='auth-challenge.js' type='module'></script>
+  `)], ['desktop/renderer/app.js', 'desktop/renderer/auth-challenge.js']);
+  for (const source of ['https://example.invalid/app.js', '../../escape.js', '/absolute.js']) {
+    assert.throws(() => controlModuleEntrypoints(`<script type="module" src="${source}"></script>`));
+  }
+});
 
 test('syntax gate argument and tracked-path schemas are exact', () => {
   assert.deepEqual(parseArguments([]), { tree: 'HEAD' });
@@ -15,6 +28,7 @@ test('syntax gate argument and tracked-path schemas are exact', () => {
   assert.throws(() => parseArguments(['--unknown']), /usage/u);
   assert.throws(() => parseArguments(['--tree', '--bad']), /invalid/u);
   assert.equal(safeTrackedPath('desktop/lib/example.js'), 'desktop/lib/example.js');
+  assert.equal(safeTrackedPath('desktop/renderer/model.mjs'), 'desktop/renderer/model.mjs');
   for (const value of ['/absolute.js', '../escape.js', 'x/../escape.js', 'x\\bad.js',
     'x:bad.js', 'x\nbad.js']) {
     assert.throws(() => safeTrackedPath(value), /invalid/u);
@@ -34,6 +48,7 @@ test('syntax gate fails closed on an empty or malformed Git enumeration', () => 
 test('syntax gate excludes generated dependency/output trees and retains exact tracked sources', () => {
   const stdout = Buffer.from([
     'desktop/main.js',
+    'desktop/renderer/model.mjs',
     'desktop/lib/space name.js',
     'desktop/node_modules/ignored.js',
     'desktop/release/ignored.js',
@@ -43,7 +58,7 @@ test('syntax gate excludes generated dependency/output trees and retains exact t
     repoRoot: '/repo', tree: 'HEAD',
     execute: () => ({ status: 0, stdout }),
   });
-  assert.deepEqual(files, ['desktop/lib/space name.js', 'desktop/main.js']);
+  assert.deepEqual(files, ['desktop/lib/space name.js', 'desktop/main.js', 'desktop/renderer/model.mjs']);
 });
 
 test('syntax gate invokes the real Node parser and rejects invalid JavaScript', () => {
@@ -52,4 +67,6 @@ test('syntax gate invokes the real Node parser and rejects invalid JavaScript', 
   const invalid = checkJavaScriptSource(Buffer.from("'use strict';\nconst answer =\n"));
   assert.equal(invalid.ok, false);
   assert.match(invalid.diagnostic, /SyntaxError/u);
+  assert.equal(checkJavaScriptSource(Buffer.from('export const answer = 42;'), { module: true }).ok, true);
+  assert.equal(checkJavaScriptSource(Buffer.from('return 42;'), { module: true }).ok, false);
 });
